chore(db): promote decodeRows helper repo-wide and consolidate 4 callsites
PR #431 introduced decodeRows<T>/decodeRowCount in lib/firms/decode-rows.ts
to isolate the dual-backend ({rows: T[]} vs T[]) row-unwrap cast at one
typed boundary. Move the helper to lib/db/decode-rows.ts (shared location
alongside the repo's other DB-shared modules) and apply the same swap to
the three other files Scout flagged:

- lib/notify/watch-confirmed.ts (2 callsites: findPriorRow, loadUserEmail)
- lib/ai/generate.ts (5 callsites: loadEventContext, fetchLastAoiBriefedAt,
  fetchEventSatellites, fetchNearestDetection, fetchPriorEvents)
- lib/db/aoi-repository.ts (10 callsites: createAoi, listAois, getAoiById,
  listAoisWithLatestBrief, listBriefsForAoi, getBriefByIdForUser,
  getBriefByShareToken, listAllBriefsForUser, listAllBriefsWithPayloadForUser,
  listMatchedDetectionsForAoi, listBriefsForAoiWithPayload)

No behavioural change. Net -37 lines.
