+++ lib/db/decode-rows.ts
